plugins/computer-vision/pyproject.toml
[project]
name = "computer-vision-claude-code"
version = "1.6.0"
description = "Desktop computer vision and input control for Claude Code on Windows"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "mss>=9.0.0",
    "pywin32>=306",
    "Pillow>=10.0.0",
    "winocr>=0.0.10",
    "comtypes>=1.4.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
ocr-fallback = ["pytesseract>=0.3.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
